+++ async-http-client/src/main/scala/org/http4s/client/asynchttpclient/AsyncHttpClient.scala
@@ -29,7 +29,7 @@ object AsyncHttpClient {
   val defaultConfig = new DefaultAsyncHttpClientConfig.Builder()
     .setMaxConnectionsPerHost(200)
     .setMaxConnections(400)
-    .setRequestTimeout(60000)
+    .setRequestTimeout(defaults.RequestTimeout.toMillis.toInt)
     .setThreadFactory(threadFactory(name = { i =>
       s"http4s-async-http-client-worker-${i}"
     }))