+++ server/src/main/scala/org/http4s/server/middleware/RequestLogger.scala
@@ -35,7 +35,8 @@ object RequestLogger {
     })(identity)
     Kleisli { req =>
       if (!logBody) {
-        def logAct = Logger.logMessage[F, Request[F]](req)(logHeaders, logBody)(log)
+        def logAct =
+          Logger.logMessage[F, Request[F]](req)(logHeaders, logBody, redactHeadersWhen)(log)
         // This construction will log on Any Error/Cancellation
         // The Completed Case is Unit, as we rely on the semantics of G
         // As None Is Successful, but we oly want to log on Some