+++ server/src/main/scala/org/http4s/server/package.scala
@@ -16,20 +16,24 @@ import java.net.{InetAddress, InetSocketAddress}
 package object server {
 
   object defaults {
-    val AsyncTimeout: Duration = 30.seconds
     val Banner =
       """|  _   _   _        _ _
          | | |_| |_| |_ _ __| | | ___
          | | ' \  _|  _| '_ \_  _(_-<
          | |_||_\__|\__| .__/ |_|/__/
          |             |_|""".stripMargin.split("\n").toList
+
     val Host = InetAddress.getLoopbackAddress.getHostAddress
     val HttpPort = 8080
-    val IdleTimeout: Duration = 30.seconds
+    val SocketAddress = InetSocketAddress.createUnresolved(Host, HttpPort)
+
+    @deprecated("Renamed to ResponseTimeout", "0.21.0-M3")
+    def AsyncTimeout: Duration = ResponseTimeout
+    val ResponseTimeout: Duration = 30.seconds
+    val IdleTimeout: Duration = 60.seconds
 
     /** The time to wait for a graceful shutdown */
     val ShutdownTimeout: Duration = 30.seconds
-    val SocketAddress = InetSocketAddress.createUnresolved(Host, HttpPort)
   }
 
   object ServerRequestKeys {
@@ -104,10 +108,10 @@ package object server {
 
     def apply[F[_], Err, T](
         authUser: Kleisli[F, Request[F], Either[Err, T]],
-        onFailure: AuthedService[Err, F]
+        onFailure: AuthedRoutes[Err, F]
     )(implicit F: Monad[F], C: Choice[Kleisli[OptionT[F, ?], ?, ?]]): AuthMiddleware[F, T] = {
-      service: AuthedService[T, F] =>
-        C.choice(onFailure, service)
+      routes: AuthedRoutes[T, F] =>
+        C.choice(onFailure, routes)
           .local { authed: AuthedRequest[F, Either[Err, T]] =>
             authed.authInfo.bimap(
               err => AuthedRequest(err, authed.req),